backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
pandas>=2.1.0
python-multipart>=0.0.6
supabase>=2.0.0
httpx>=0.25.0
